% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_final_processing.R
\name{harmonize_platform_fields}
\alias{harmonize_platform_fields}
\title{Harmonize platform-specific fields based on \code{ngs_platform}}
\usage{
harmonize_platform_fields(df)
}
\arguments{
\item{df}{Data frame with ngs_platform + platform-specific cols}
}
\value{
Data frame with harmonized platform fields
}
\description{
Nanopore → blank illumina_platform
Illumina → blank nanopore_platform
}
